chore: move investor into lib (#4508)
* chore: move investor into lib

* chore: remove path
