--- packages/investor/.npmignore
@@ -1,5 +1,0 @@
-coverage
-jest.config.js
-tsconfig.json
-*.test.js
-*.test.d.ts
