--- packages/investor/CHANGELOG.md
@@ -1,72 +1,0 @@
-# [@shapeshiftoss/investor-v3.1.3](https://github.com/shapeshift/lib/compare/@shapeshiftoss/investor-v3.1.2...@shapeshiftoss/investor-v3.1.3) (2023-03-03)
-
-
-### Bug Fixes
-
-* **common:** osmosis keepkey swaps ([#1221](https://github.com/shapeshift/lib/issues/1221)) ([156f2ac](https://github.com/shapeshift/lib/commit/156f2acf815fe05a74fdfbe7238f813936f2a2b9))
-
-# [@shapeshiftoss/investor-v3.1.2](https://github.com/shapeshift/lib/compare/@shapeshiftoss/investor-v3.1.1...@shapeshiftoss/investor-v3.1.2) (2023-02-24)
-
-# [@shapeshiftoss/investor-v3.1.1](https://github.com/shapeshift/lib/compare/@shapeshiftoss/investor-v3.1.0...@shapeshiftoss/investor-v3.1.1) (2023-02-23)
-
-
-### Bug Fixes
-
-* **chain-adapters:** use snake_case for all osmosis tx fields ([#1205](https://github.com/shapeshift/lib/issues/1205)) ([957569a](https://github.com/shapeshift/lib/commit/957569a5e3460ae56a40d6d8337398504f741a8f))
-
-# [@shapeshiftoss/investor-v3.1.0](https://github.com/shapeshift/lib/compare/@shapeshiftoss/investor-v3.0.1...@shapeshiftoss/investor-v3.1.0) (2023-01-13)
-
-
-### Features
-
-* add optimism ([#1157](https://github.com/shapeshift/lib/issues/1157)) ([177725d](https://github.com/shapeshift/lib/commit/177725dc04c00ba982bbebe8e84a7edcff395367))
-
-# [@shapeshiftoss/investor-v3.0.1](https://github.com/shapeshift/lib/compare/@shapeshiftoss/investor-v3.0.0...@shapeshiftoss/investor-v3.0.1) (2022-10-19)
-
-# [@shapeshiftoss/investor-v3.0.0](https://github.com/shapeshift/lib/compare/@shapeshiftoss/investor-v2.1.1...@shapeshiftoss/investor-v3.0.0) (2022-10-13)
-
-
-### Features
-
-* **investor:** enforce `bip44Params` on `signAndBroadcast` ([#1061](https://github.com/shapeshift/lib/issues/1061)) ([19a6c7c](https://github.com/shapeshift/lib/commit/19a6c7c732db506b69548791c188868104738115))
-
-
-### BREAKING CHANGES
-
-* **investor:** makes bip44Params required for signAndBroadcast
-
-Co-authored-by: gomes <17035424+gomesalexandre@users.noreply.github.com>
-
-# [@shapeshiftoss/investor-v2.1.1](https://github.com/shapeshift/lib/compare/@shapeshiftoss/investor-v2.1.0...@shapeshiftoss/investor-v2.1.1) (2022-09-16)
-
-# [@shapeshiftoss/investor-v2.1.0](https://github.com/shapeshift/lib/compare/@shapeshiftoss/investor-v2.0.1...@shapeshiftoss/investor-v2.1.0) (2022-09-12)
-
-
-### Features
-
-* add THORChain support ([#1010](https://github.com/shapeshift/lib/issues/1010)) ([d7c3b72](https://github.com/shapeshift/lib/commit/d7c3b72bbda9795f87fa8f73c35926c95026a3c2))
-
-# [@shapeshiftoss/investor-v2.0.1](https://github.com/shapeshift/lib/compare/@shapeshiftoss/investor-v2.0.0...@shapeshiftoss/investor-v2.0.1) (2022-09-09)
-
-# [@shapeshiftoss/investor-v2.0.0](https://github.com/shapeshift/lib/compare/@shapeshiftoss/investor-v1.0.2...@shapeshiftoss/investor-v2.0.0) (2022-09-08)
-
-
-### Features
-
-* **investor-yearn:** accept bip44Params in signAndBroadcast input ([#1014](https://github.com/shapeshift/lib/issues/1014)) ([dc65033](https://github.com/shapeshift/lib/commit/dc65033851000a476db4f3ac073d26c40094969c))
-
-
-### BREAKING CHANGES
-
-* **investor-yearn:** signAndBroadcast now accepts an optional bip44Params property and will throw if not passed.
-
-# [@shapeshiftoss/investor-v1.0.2](https://github.com/shapeshift/lib/compare/@shapeshiftoss/investor-v1.0.1...@shapeshiftoss/investor-v1.0.2) (2022-08-15)
-
-# [@shapeshiftoss/investor-v1.0.1](https://github.com/shapeshift/lib/compare/@shapeshiftoss/investor-v1.0.0...@shapeshiftoss/investor-v1.0.1) (2022-06-08)
-
-# @shapeshiftoss/investor-v1.0.0 (2022-06-02)
-
-
-### Features
-
-* **investor:** add investor package ([#720](https://github.com/shapeshift/lib/issues/720)) ([bc3cc62](https://github.com/shapeshift/lib/commit/bc3cc62b6dfad754da27945ae07be17b6fc04124))
