--- packages/investor/README.md
@@ -1,3 +1,0 @@
-# @shapeshiftoss/investor
-
-ShapeShift's single sided staking abstraction.
