--- packages/investor/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "@shapeshiftoss/investor",
-  "version": "3.1.3",
-  "description": "A core abstraction for single-sided staking plugins",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "directories": {
-    "lib": "dist",
-    "src": "src"
-  },
-  "files": [
-    "dist"
-  ],
-  "publishConfig": {
-    "access": "public"
-  },
-  "scripts": {
-    "build": "yarn clean && yarn run -T tsc --build",
-    "clean": "rm -rf dist tsconfig.tsbuildinfo",
-    "dev": "yarn run -T tsc --build --watch"
-  },
-  "dependencies": {
-    "@shapeshiftoss/types": "workspace:^"
-  }
-}
