--- packages/investor/src/index.ts
@@ -1,3 +1,0 @@
-export * from './Investor'
-export * from './InvestorOpportunity'
-export * from './Extensions'
