--- packages/investor/tsconfig.json
@@ -1,12 +1,0 @@
-{
-  "extends": "../../tsconfig.packages.json",
-  "compilerOptions": {
-    "rootDir": "src",
-    "outDir": "dist"
-  },
-  "include": ["src/**/*"],
-  "exclude": ["dist"],
-  "references": [
-    { "path": "../types" },
-  ]
-}
