+++ src/lib/investor/index.ts
@@ -1,0 +1,3 @@
+export * from './types/Investor'
+export * from './types/InvestorOpportunity'
+export * from './types/Extensions'
