+++ src/lib/investor/investor-idle/IdleInvestor.ts
@@ -1,5 +1,4 @@
 import type { ChainAdapter } from '@shapeshiftoss/chain-adapters'
-import type { Investor } from '@shapeshiftoss/investor'
 import type { Logger } from '@shapeshiftoss/logger'
 import type { KnownChainIds } from '@shapeshiftoss/types'
 // import https from 'https'
@@ -7,6 +6,7 @@ import { find } from 'lodash'
 import filter from 'lodash/filter'
 import Web3 from 'web3'
 import type { Contract } from 'web3-eth-contract'
+import type { Investor } from 'lib/investor'
 import type { IdleVault } from 'lib/investor/investor-idle/constants'
 import { ssRouterAbi, ssRouterContractAddress } from 'lib/investor/investor-idle/constants'
 import type { PreparedTransaction } from 'lib/investor/investor-idle/IdleOpportunity'
