+++ src/lib/investor/investor-yearn/YearnInvestor.ts
@@ -1,5 +1,4 @@
 import type { ChainAdapter } from '@shapeshiftoss/chain-adapters'
-import type { Investor } from '@shapeshiftoss/investor'
 import type { KnownChainIds } from '@shapeshiftoss/types'
 import { type ChainId, type VaultMetadata, Yearn } from '@yfi/sdk'
 import { ethers } from 'ethers'
@@ -7,6 +6,7 @@ import { find } from 'lodash'
 import filter from 'lodash/filter'
 import Web3 from 'web3'
 import type { Contract } from 'web3-eth-contract'
+import type { Investor } from 'lib/investor'
 
 import { ssRouterAbi, ssRouterContractAddress } from './constants'
 import type { PreparedTransaction } from './YearnOpportunity'
