+++ src/lib/investor/investor-yearn/YearnOpportunity.ts
@@ -2,12 +2,6 @@ import { toAssetId } from '@shapeshiftoss/caip'
 import type { ChainAdapter } from '@shapeshiftoss/chain-adapters'
 import { toAddressNList } from '@shapeshiftoss/chain-adapters'
 import type { ETHSignTx, HDWallet } from '@shapeshiftoss/hdwallet-core'
-import type {
-  ApprovalRequired,
-  DepositWithdrawArgs,
-  FeePriority,
-  InvestorOpportunity,
-} from '@shapeshiftoss/investor'
 import { Logger } from '@shapeshiftoss/logger'
 import type { BIP44Params, KnownChainIds } from '@shapeshiftoss/types'
 import type { Yearn } from '@yfi/sdk'
@@ -18,6 +12,12 @@ import toLower from 'lodash/toLower'
 import type Web3 from 'web3'
 import type { Contract } from 'web3-eth-contract'
 import { numberToHex } from 'web3-utils'
+import type {
+  ApprovalRequired,
+  DepositWithdrawArgs,
+  FeePriority,
+  InvestorOpportunity,
+} from 'lib/investor'
 
 import { erc20Abi, MAX_ALLOWANCE, ssRouterContractAddress, yv2VaultAbi } from './constants'
 import { bn, bnOrZero } from './utils'
