+++ src/lib/investor/types/Extensions.ts
